requirements.txt
anyio==3.6.2
arrow==1.2.3
bcrypt==4.0.1
binaryornot==0.4.4
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cookiecutter==1.7.3
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.95.1
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jinja2-time==0.2.0
manage-fastapi==1.1.1
MarkupSafe==2.1.2
orjson==3.8.11
passlib==1.7.4
poyo==0.5.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
pydantic==1.10.7
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
python-slugify==8.0.1
PyYAML==6.0
questionary==1.10.0
requests==2.29.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.12
starlette==0.26.1
text-unidecode==1.3
typer==0.4.2
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.2
